+++ src/commands/runners/CollectionBuilder.js
@@ -2,18 +2,19 @@
 import { mkdir, rm } from 'node:fs/promises'
 import { join } from 'node:path'
 
-// External Packages
-import merge from 'deepmerge'
-
 // Internal Imports
 import BuildFactory from '../../factory/BuildFactory.js'
+import Session from '../../models/static/Session.js'
+import LocalesProcessor from '../../processors/LocalesProcessor.js'
+import { validateExternalLocation } from '../../utils/ExternalComponentUtils.js'
 import FileUtils from '../../utils/FileUtils.js'
+import { exitWithError } from '../../utils/NodeUtils.js'
 import WebUtils from '../../utils/WebUtils.js'
 import JavaScriptProcessor from '../../processors/JavaScriptProcessor.js'
 import LocaleUtils from '../../utils/LocaleUtils.js'
 import StylesProcessor from '../../processors/StylesProcessor.js'
-import Timer from '../../utils/Timer.js'
-import logger, { logChildItem } from '../../utils/Logger.js'
+import { getTimeElapsed, getTimer } from '../../utils/Timer.js'
+import { logChildItem, WARN_LOG_LEVEL } from '../../utils/Logger.js'
 
 class CollectionBuilder {
   /**
@@ -22,39 +23,71 @@ class CollectionBuilder {
    * @return {Promise<module:models/Collection>}
    */
   static async build (collection) {
-    const allComponents = [...collection.components, ...collection.snippets]
+    const allComponents = collection.allComponents
 
-    // Create Collection Build model and reset folders
-    const buildCollectionTimer = Timer.getTimer()
+    const buildCollectionTimer = getTimer()
     collection.build = BuildFactory.fromCollection(collection)
-    await this.#resetBuildFolders(collection)
-    logChildItem(`Collection Build Initialized (${Timer.getEndTimerInSeconds(buildCollectionTimer)} seconds)`)
+    await this.#resetBuildFolders(collection.build)
+    logChildItem(`Collection Build Initialized (${getTimeElapsed(buildCollectionTimer)} seconds)`);
 
-    // Gather and build Stylesheets
-    const mainStylesheets = this.getMainStylesheets(allComponents)
-
-    if (mainStylesheets.length) {
-      const buildStylesTimer = Timer.getTimer()
-      collection.build.styles = await StylesProcessor.buildStylesBundle(mainStylesheets, collection.build.stylesheet, collection.rootFolder)
-      logChildItem(`Styles Ready (${Timer.getEndTimerInSeconds(buildStylesTimer)} seconds)`)
-    }
-    // Build Collection JS Files
-    const jsFiles = this.getJsFiles(allComponents)
-
-    if (jsFiles.length) {
-      const buildScriptsTimer = Timer.getTimer()
-      collection.importMapEntries = await JavaScriptProcessor.buildJavaScript(jsFiles, collection.build.importMapFile, collection.rootFolder)
-      logChildItem(`Scripts Ready (${Timer.getEndTimerInSeconds(buildScriptsTimer)} seconds)`)
-    } else {
-      logger.warn('No Javascript Files Found. Javascript Build Process Was Skipped.')
-    }
-
-    // Build Locales
-    const buildLocalesTimer = Timer.getTimer()
-    collection.build.locales = this.buildLocales(collection.components)
-    logChildItem(`Locales Ready (${Timer.getEndTimerInSeconds(buildLocalesTimer)} seconds)`)
+    [collection.importMapEntries, collection.build.locales, collection.build.styles] = await Promise.all([
+      this.#buildJavaScript(allComponents, collection.build.importMapFile, collection.rootFolder),
+      this.#buildLocales(allComponents, collection.rootFolder),
+      this.#buildStyles(allComponents, collection.build.stylesheet, collection.rootFolder)
+    ])
 
     return collection
+  }
+
+  /**
+   * Builds JavaScript files for the given collection and components.
+   *
+   * @param {(Component|Snippet)[]} components - An array of all components.
+   * @param {string} importMapFile - The collection's import map file.
+   * @param {string} cwd - The working directory.
+   * @return {Map<string, string>} - A promise that resolves when the JavaScript files have been built.
+   */
+  static async #buildJavaScript (components, importMapFile, cwd) {
+    const jsFiles = this.#getJsFiles(components)
+
+    if (jsFiles.length) {
+      return JavaScriptProcessor.buildJavaScript(jsFiles, importMapFile, cwd)
+    } else {
+      logChildItem('No Javascript Files Found. Javascript Build Process Was Skipped.', WARN_LOG_LEVEL)
+    }
+  }
+
+  /**
+   * Builds locales for the given collection and all components.
+   *
+   * @param {(Component|Snippet)[]} components - All components to extract liquid code from.
+   * @param {string} cwd - The working directory.
+   * @return {Promise<{}>} - A promise that resolves when the locales are built.
+   */
+  static async #buildLocales (components, cwd) {
+    const liquidCodeElements = components.map(component => component.liquidCode)
+    try {
+      const validLocalesRepo = await validateExternalLocation(Session.localesRepo, cwd)
+      return LocalesProcessor.build(liquidCodeElements, validLocalesRepo, cwd)
+    } catch (error) {
+      exitWithError('Source Locales DB Folder or Repository is invalid: ' + error.message)
+    }
+  }
+
+  /**
+   * Builds the styles for the given collection and all components.
+   *
+   * @param {(Component|Snippet)[]} components - An array containing all the components.
+   * @param {string} outputFile - The collection's css bundle file.
+   * @param {string} cwd - The working directory.
+   * @return {Promise<string>} - A promise that resolves when the styles are built.
+   */
+  static async #buildStyles (components, outputFile, cwd) {
+    const mainStylesheets = this.#getMainStylesheets(components)
+
+    if (mainStylesheets.length) {
+      return StylesProcessor.buildStylesBundle(mainStylesheets, outputFile, cwd)
+    }
   }
 
   /**
@@ -63,28 +96,24 @@ class CollectionBuilder {
    * @returns {Promise<Awaited<unknown>[]>}
    */
   static async deployToBuildFolder (collection) {
-    const allComponents = [...collection.components, ...collection.snippets]
-    const allSnippets = [...collection.components, ...collection.snippets]
+    const allComponents = collection.allComponents
 
     const localesWritePromise = LocaleUtils.writeLocales(collection.build.locales, collection.build.localesFolder)
+    const snippetFilesWritePromises = allComponents.map(component =>
+      FileUtils.saveFile(join(collection.build.snippetsFolder, `${component.name}.liquid`), component.build.liquidCode))
 
-    // Write Component Liquid Files
-    const snippetFilesWritePromises = allSnippets.map(component =>
-      FileUtils.writeFile(join(collection.build.snippetsFolder, `${component.name}.liquid`), component.build.liquidCode))
-
-    // Gather & Copy Assets Files
-    const allAssetFiles = this.getAssetFiles(allComponents)
+    const allAssetFiles = this.#getAssetFiles(allComponents)
     const copyAssetsPromise = FileUtils.copyFilesToFolder(allAssetFiles, collection.build.assetsFolder)
 
     const promises = [
-      FileUtils.writeFile(collection.build.stylesheet, collection.build.styles),
+      FileUtils.saveFile(collection.build.stylesheet, collection.build.styles),
       localesWritePromise,
       ...snippetFilesWritePromises,
       copyAssetsPromise
     ]
 
     if (collection.importMapEntries?.size) {
-      promises.push(this.deployImportMapFiles(collection.importMapEntries, collection.build.assetsFolder))
+      promises.push(this.#deployImportMapFiles(collection.importMapEntries, collection.build.assetsFolder))
     }
 
     return Promise.all(promises)
@@ -95,7 +124,7 @@ class CollectionBuilder {
    * @param {Map<string, string>} buildEntries
    * @param {string} assetsFolder
    */
-  static async deployImportMapFiles (buildEntries, assetsFolder) {
+  static async #deployImportMapFiles (buildEntries, assetsFolder) {
     const localFiles = []
     const remoteFiles = []
     for (const [, modulePath] of buildEntries) {
@@ -109,29 +138,15 @@ class CollectionBuilder {
   }
 
   /**
-   * Build Collection Storefront Locales
-   * @param {(Component|Snippet)[]} components
-   * @return {Object}
-   */
-  static buildLocales (components) {
-    let buildLocales = {}
-
-    for (const component of components) {
-      buildLocales = merge(buildLocales, component.build.locales)
-    }
-
-    return buildLocales
-  }
-
-  /**
+   * Retrieves asset files from an array of Component or Snippet objects.
    *
-   * @param {(Component|Snippet)[]} components
-   * @return {string[]}
+   * @param {(Component|Snippet)[]} components - An array of components or snippets.
+   * @return {string[]} - An array of asset files.
    */
-  static getAssetFiles (components) {
-    const filteredComponents = components.filter(component => component.files.assetFiles?.length)
-
-    return (filteredComponents.map(component => component.files.assetFiles)).flat()
+  static #getAssetFiles (components) {
+    return components
+      .filter(component => component.files.assetFiles?.length)
+      .flatMap(component => component.files.assetFiles)
   }
 
   /**
@@ -139,9 +154,10 @@ class CollectionBuilder {
    * @param {(Component|Snippet)[]} components
    * @return {string[]}
    */
-  static getJsFiles (components) {
-    const componentsWithScriptFiles = components.filter(component => component.files.javascriptIndex)
-    return componentsWithScriptFiles.map(component => component.files.javascriptIndex)
+  static #getJsFiles (components) {
+    return components
+      .filter(component => component.files.javascriptIndex)
+      .map(component => component.files.javascriptIndex)
   }
 
   /**
@@ -149,26 +165,27 @@ class CollectionBuilder {
    * @param {(Component|Snippet)[]} components
    * @return {string[]}
    */
-  static getMainStylesheets (components) {
-    const componentsWithStylesheets = components.filter(component => component.files.mainStylesheet)
-    return componentsWithStylesheets.map(component => component.files.mainStylesheet)
+  static #getMainStylesheets (components) {
+    return components
+      .filter(component => component.files.mainStylesheet)
+      .map(component => component.files.mainStylesheet)
   }
 
   /**
    * Reset Collection Build Folders
-   * @param {module:models/Collection} collection
+   * @param {CollectionBuild} build
    * @return {Promise<Awaited<unknown>[]>}
    */
-  static async #resetBuildFolders (collection) {
-    await rm(collection.build.rootFolder, { force: true, recursive: true })
-    await mkdir(collection.build.rootFolder, { recursive: true })
+  static async #resetBuildFolders (build) {
+    await rm(build.rootFolder, { force: true, recursive: true })
+    await mkdir(build.rootFolder, { recursive: true })
 
     const buildFolders = [
-      collection.build.assetsFolder,
-      collection.build.configFolder,
-      collection.build.localesFolder,
-      collection.build.sectionsFolder,
-      collection.build.snippetsFolder
+      build.assetsFolder,
+      build.configFolder,
+      build.localesFolder,
+      build.sectionsFolder,
+      build.snippetsFolder
     ]
     const mkdirPromises = buildFolders.map(buildFolder => mkdir(buildFolder, { recursive: true }))